requirements.txt
homeassistant>=2023.7.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
voluptuous>=0.13.0
pyserial>=3.5 